requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==36.0.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.10
django-bootstrap-v5==1.0.8
django-bootstrap5==21.3
django-braces==1.15.0
django-oauth-toolkit==1.6.1
django-rest-framework-social-oauth2==1.1.0
django-rest-knox==4.1.0
djangorestframework==3.13.1
idna==3.3
importlib-metadata==3.10.1
jwcrypto==1.0
oauthlib==3.1.1
Pillow==8.4.0
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
rest-social-auth==7.0.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
soupsieve==2.3.1
sqlparse==0.4.2
urllib3==1.26.7
wrapt==1.13.3
zipp==3.7.0
